# src/app/api/v1/users/route.ts
import { users } from "@/services/UserService";
import { NextRequest, NextResponse } from "next/server";

export async function GET() {
  return NextResponse.json(users);
}

export async function POST(request: NextRequest) {
  const body = await request.json();
  const { name, email } = body;

  if (!name || !email) {
    return NextResponse.json(
      { error: "Name and email are required" },
      { status: 400 }
    );
  }

  const newUser = {
    id: users.length > 0 ? users[users.length - 1].id + 1 : 1,
    name,
    email,
  };

  users.push(newUser);

  return NextResponse.json(newUser, { status: 201 });
}